# Fixing  KeyError: \'BGP_NEIGHBOR\' for BGP performance and scalability script