<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary: Fixing  KeyError: \'BGP_NEIGHBOR\' for BGP performance and scalability script
Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] Test case(new/improvement)


### Back port request
- [ ] 201911
- [ ] 202012
- [ ] 202205
- [ ] 202305
- [ ] 202311

### Approach
#### What is the motivation for this PR?
Fixing  KeyError: \'BGP_NEIGHBOR\' for BGP performance and scalability script
#### How did you do it?
Modified BGP configuration to use vtysh commands instead of json dump
#### How did you verify/test it?
Verified on Arista DUT
#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->
### Output:

AzDevOps@8e6233df64b9:~/sonic-mgmt/tests$ py.test --inventory ../ansible/snappi-sonic --host-pattern sonic-s6100-dut --testbed vms-snappi-sonic --testbed_file ../ansible/testbed.csv --show-capture=stdout --log-cli-level info --showlocals -ra --allow_recover --skip_sanity --disable_loganalyzer snappi_tests/bgp/test_bgp_scalability.py 
========================================================================================================================= test session starts =========================================================================================================================
platform linux -- Python 3.8.10, pytest-7.4.0, pluggy-1.3.0
ansible: 2.13.13
rootdir: /var/AzDevOps/sonic-mgmt/tests
configfile: pytest.ini
plugins: allure-pytest-2.8.22, ansible-4.0.0, forked-1.6.0, html-4.1.0, metadata-3.0.0, repeat-0.9.3, xdist-1.28.0
collecting ... 
------------------------------------------------------------------------------------------------------------------------- live log collection -------------------------------------------------------------------------------------------------------------------------
19:17:06 __init__.pytest_collection_modifyitems   L0581 INFO   | Available basic facts that can be used in conditional skip:
{
  "topo_type": "ptf",
  "topo_name": "ptf64",
  "testbed": "vms-snappi-sonic"
}
collected 2 items                                                                                                                                                                                                                                                     

snappi_tests/bgp/test_bgp_scalability.py::test_duthost_bgp_scalability_config[1] 
--------------------------------------------------------------------------------------------------------------------------- live log setup ----------------------------------------------------------------------------------------------------------------------------
19:17:06 __init__.set_default                     L0053 INFO   | Completeness level not set during test execution. Setting to default level: CompletenessLevel.basic
19:17:06 __init__.check_test_completeness         L0152 INFO   | Test has no defined levels. Continue without test completeness checks
19:17:12 ptfhost_utils.run_icmp_responder_session L0239 INFO   | Skip running icmp_responder at session level, it is only for dualtor testbed with active-active mux ports.
19:17:12 conftest.creds_on_dut                    L0723 INFO   | dut sonic-s6100-dut belongs to groups ['snappi-sonic', 'sonic', 'sonic_dell64_40', 'fanout']
19:17:12 conftest.creds_on_dut                    L0747 INFO   | skip empty var file ../ansible/group_vars/all/corefile_uploader.yml
19:17:12 conftest.creds_on_dut                    L0747 INFO   | skip empty var file ../ansible/group_vars/all/env.yml
19:17:12 conftest.nbrhosts                        L0540 INFO   | No VMs exist for this topology: ptf64
19:17:12 conftest.core_dump_and_config_check      L2052 INFO   | Dumping Disk and Memory Space informataion before test on sonic-s6100-dut
19:17:14 conftest.core_dump_and_config_check      L2056 INFO   | Collecting core dumps before test on sonic-s6100-dut
19:17:14 conftest.core_dump_and_config_check      L2065 INFO   | Collecting running config before test on sonic-s6100-dut
19:17:20 __init__.sanity_check                    L0125 INFO   | Skip sanity check according to command line argument
19:17:20 conftest.generate_params_dut_hostname    L1127 INFO   | Using DUTs ['sonic-s6100-dut'] in testbed 'vms-snappi-sonic'
19:17:20 conftest.rand_one_dut_hostname           L0397 INFO   | Randomly select dut sonic-s6100-dut for testing
19:17:20 __init__._fixture_generator_decorator    L0081 INFO   | -------------------- fixture enable_packet_aging_after_test setup starts --------------------
19:17:20 __init__._fixture_generator_decorator    L0085 INFO   | -------------------- fixture enable_packet_aging_after_test setup ends --------------------
19:17:20 __init__._fixture_generator_decorator    L0081 INFO   | -------------------- fixture rand_lossless_prio setup starts --------------------
19:17:20 __init__._fixture_generator_decorator    L0085 INFO   | -------------------- fixture rand_lossless_prio setup ends --------------------
19:17:20 __init__._fixture_generator_decorator    L0081 INFO   | -------------------- fixture rand_lossy_prio setup starts --------------------
19:17:20 __init__._fixture_generator_decorator    L0085 INFO   | -------------------- fixture rand_lossy_prio setup ends --------------------
19:17:20 __init__._fixture_generator_decorator    L0081 INFO   | -------------------- fixture start_pfcwd_after_test setup starts --------------------
19:17:20 __init__._fixture_generator_decorator    L0085 INFO   | -------------------- fixture start_pfcwd_after_test setup ends --------------------
19:17:20 __init__._fixture_func_decorator         L0069 INFO   | -------------------- fixture tgen_ports setup starts --------------------
19:17:21 __init__._fixture_func_decorator         L0076 INFO   | -------------------- fixture tgen_ports setup ends --------------------
19:17:21 __init__.loganalyzer                     L0051 INFO   | Log analyzer is disabled
---------------------------------------------------------------------------------------------------------------------------- live log call ----------------------------------------------------------------------------------------------------------------------------
19:17:28 bgp_test_gap_helper.duthost_bgp_scalabil L0244 INFO   | Removing configured IP and IPv6 Address from Ethernet0
19:17:31 bgp_test_gap_helper.duthost_bgp_scalabil L0244 INFO   | Removing configured IP and IPv6 Address from Ethernet4
19:17:34 bgp_test_gap_helper.duthost_bgp_scalabil L0257 INFO   | Configuring Ethernet0 to PortChannel1
19:17:38 bgp_test_gap_helper.duthost_bgp_scalabil L0257 INFO   | Configuring Ethernet4 to PortChannel2
19:17:43 bgp_test_gap_helper.duthost_bgp_scalabil L0316 INFO   | Configuring BGP v4 Neighbor 22.1.1.2, v6 Neighbor 2000:2::e
PASSED                                                                                                                                                                                                                                                          [ 50%]
snappi_tests/bgp/test_bgp_scalability.py::test_bgp_scalability_16k_v4_routes[64-1-16000-1] 
--------------------------------------------------------------------------------------------------------------------------- live log setup ----------------------------------------------------------------------------------------------------------------------------
19:17:44 __init__.set_default                     L0053 INFO   | Completeness level not set during test execution. Setting to default level: CompletenessLevel.basic
19:17:44 __init__.check_test_completeness         L0152 INFO   | Test has no defined levels. Continue without test completeness checks
19:17:44 __init__._fixture_func_decorator         L0069 INFO   | -------------------- fixture snappi_api_serv_ip setup starts --------------------
19:17:44 __init__._fixture_func_decorator         L0076 INFO   | -------------------- fixture snappi_api_serv_ip setup ends --------------------
19:17:44 __init__._fixture_func_decorator         L0069 INFO   | -------------------- fixture snappi_api_serv_port setup starts --------------------
19:17:44 __init__._fixture_func_decorator         L0076 INFO   | -------------------- fixture snappi_api_serv_port setup ends --------------------
19:17:44 __init__._fixture_generator_decorator    L0081 INFO   | -------------------- fixture cvg_api setup starts --------------------
19:17:44 __init__._fixture_generator_decorator    L0085 INFO   | -------------------- fixture cvg_api setup ends --------------------
19:17:44 __init__.loganalyzer                     L0051 INFO   | Log analyzer is disabled
---------------------------------------------------------------------------------------------------------------------------- live log call ----------------------------------------------------------------------------------------------------------------------------
19:17:44 connection._warn                         L0246 WARNING| Verification of certificates is disabled
19:17:44 connection._info                         L0243 INFO   | Determining the platform and rest_port using the 10.36.77.63 address...
19:17:44 connection._warn                         L0246 WARNING| Unable to connect to http://10.36.77.63:11009.
19:17:44 connection._info                         L0243 INFO   | Connection established to `https://10.36.77.63:11009 on windows`
19:17:44 connection._info                         L0243 INFO   | Using IxNetwork api server version 9.30.2212.5
19:17:44 connection._info                         L0243 INFO   | User info IxNetwork/ST-REG-VM/8010
19:17:44 snappi_api.info                          L1132 INFO   | snappi-0.9.1
19:17:44 snappi_api.info                          L1132 INFO   | snappi_ixnetwork-0.8.2
19:17:44 snappi_api.info                          L1132 INFO   | ixnetwork_restpy-1.0.64
19:17:44 snappi_api.info                          L1132 INFO   | Config validation 0.007s
19:17:45 snappi_api.info                          L1132 INFO   | Ports configuration 0.048s
19:17:45 snappi_api.info                          L1132 INFO   | Captures configuration 0.029s
19:17:51 snappi_api.info                          L1132 INFO   | Location hosts ready [10.36.78.53] 2.056s
19:17:52 snappi_api.info                          L1132 INFO   | Speed conversion is not require for (port.name, speed) : [('Source', 'novusHundredGigNonFanOut'), ('Destination', 'novusHundredGigNonFanOut')]
19:17:52 snappi_api.info                          L1132 INFO   | Aggregation mode speed change 0.306s
19:17:52 snappi_api.info                          L1132 INFO   | Location configuration 6.508s
19:17:52 snappi_api.info                          L1132 INFO   | Layer1 configuration 0.056s
19:17:53 snappi_api.info                          L1132 INFO   | Lag Configuration 1.052s
19:17:53 snappi_api.info                          L1132 INFO   | Lag Ethernet Configuration 0.175s
19:17:53 snappi_api.info                          L1132 INFO   | Lag Protocol Configuration 0.639s
19:17:54 snappi_api.info                          L1132 INFO   | Convert device config : 0.153s
19:17:54 snappi_api.info                          L1132 INFO   | Create IxNetwork device config : 0.001s
19:17:55 snappi_api.info                          L1132 INFO   | Push IxNetwork device config : 1.566s
19:17:55 snappi_api.info                          L1132 INFO   | Devices configuration 1.728s
19:17:56 snappi_api.info                          L1132 INFO   | Flows configuration 0.937s
19:17:57 snappi_api.info                          L1132 INFO   | Start interfaces 0.491s
19:18:01 bgp_test_gap_helper.run_traffic          L0684 INFO   | Starting all protocols ...
19:18:16 snappi_api.info                          L1132 INFO   | Setting protocol state 15.013s
19:18:16 utilities.wait                           L0104 INFO   | Pause 30 seconds, reason: For Protocols To start
19:18:46 bgp_test_gap_helper.run_traffic          L0690 INFO   | Starting Traffic
19:19:06 snappi_api.info                          L1132 INFO   | Flows generate/apply 2.869s
19:19:20 snappi_api.info                          L1132 INFO   | Flows clear statistics 13.210s
19:19:20 snappi_api.info                          L1132 INFO   | Captures start 0.000s
19:19:23 snappi_api.info                          L1132 INFO   | Flows start 3.560s
19:19:23 utilities.wait                           L0104 INFO   | Pause 50 seconds, reason: For Traffic To start and stabilize
19:20:15 bgp_test_gap_helper.run_traffic          L0692 INFO   | ROUTE_THRESHOLD_EXCEEDED FOUND in syslog!!!!!!!!
19:20:16 bgp_test_gap_helper.stop_traffic         L0713 INFO   | Stopping Traffic
19:20:21 snappi_api.info                          L1132 INFO   | Flows stop 5.266s
19:20:21 utilities.wait                           L0104 INFO   | Pause 20 seconds, reason: For Traffic To stop
19:20:41 bgp_test_gap_helper.stop_traffic         L0719 INFO   | Stopping all protocols ...
19:20:41 snappi_api.info                          L1132 INFO   | Setting protocol state 0.020s
19:20:41 utilities.wait                           L0104 INFO   | Pause 20 seconds, reason: For Protocols To STOP
19:21:02 bgp_test_gap_helper.get_bgp_scalability_ L0749 INFO   | |---- Tx Frame: 33454533 ----|
19:21:02 bgp_test_gap_helper.get_bgp_scalability_ L0750 INFO   | |---- Rx Frame: 33454533 ----|
19:21:02 bgp_test_gap_helper.get_bgp_scalability_ L0751 INFO   | |---- Loss % : 0.0 ----|
19:21:02 bgp_test_gap_helper.get_bgp_scalability_ L0758 INFO   | PASSED : No Loss observerd in traffic item and THRESHOLD_EXCEEDED warning message observed in syslog
PASSED                                                                                                                                                                                                                                                          [100%]
-------------------------------------------------------------------------------------------------------------------------- live log teardown --------------------------------------------------------------------------------------------------------------------------
19:21:02 __init__._fixture_generator_decorator    L0093 INFO   | -------------------- fixture cvg_api teardown starts --------------------
19:21:02 __init__._fixture_generator_decorator    L0102 INFO   | -------------------- fixture cvg_api teardown ends --------------------
19:21:02 __init__._fixture_generator_decorator    L0093 INFO   | -------------------- fixture start_pfcwd_after_test teardown starts --------------------
19:21:03 __init__._fixture_generator_decorator    L0102 INFO   | -------------------- fixture start_pfcwd_after_test teardown ends --------------------
19:21:03 __init__._fixture_generator_decorator    L0093 INFO   | -------------------- fixture rand_lossy_prio teardown starts --------------------
19:21:03 __init__._fixture_generator_decorator    L0102 INFO   | -------------------- fixture rand_lossy_prio teardown ends --------------------
19:21:03 __init__._fixture_generator_decorator    L0093 INFO   | -------------------- fixture rand_lossless_prio teardown starts --------------------
19:21:03 __init__._fixture_generator_decorator    L0102 INFO   | -------------------- fixture rand_lossless_prio teardown ends --------------------
19:21:03 __init__._fixture_generator_decorator    L0093 INFO   | -------------------- fixture enable_packet_aging_after_test teardown starts --------------------
19:21:03 __init__._fixture_generator_decorator    L0102 INFO   | -------------------- fixture enable_packet_aging_after_test teardown ends --------------------
19:21:03 conftest.core_dump_and_config_check      L2093 INFO   | Dumping Disk and Memory Space informataion after test on sonic-s6100-dut
19:21:05 conftest.core_dump_and_config_check      L2097 INFO   | Collecting core dumps after test on sonic-s6100-dut
19:21:05 conftest.core_dump_and_config_check      L2114 INFO   | Collecting running config after test on sonic-s6100-dut
19:21:06 conftest.core_dump_and_config_check      L2238 WARNING| Core dump or config check failed for test_bgp_scalability.py, results: {"core_dump_check": {"pass": true, "new_core_dumps": {"sonic-s6100-dut": []}}, "config_db_check": {"pass": false, "pre_only_config": {"sonic-s6100-dut": {"null": {"VLAN_MEMBER": {"Vlan2|Ethernet0": {"tagging_mode": "untagged"}, "Vlan2|Ethernet12": {"tagging_mode": "untagged"}, "Vlan2|Ethernet4": {"tagging_mode": "untagged"}, "Vlan2|Ethernet8": {"tagging_mode": "untagged"}}}}}, "cur_only_config": {"sonic-s6100-dut": {"null": {"PORTCHANNEL_INTERFACE": {"PortChannel1": {}, "PortChannel2": {}, "PortChannel1|2000:1::1/124": {}, "PortChannel1|21.1.1.1/24": {}, "PortChannel2|2000:2::1/124": {}, "PortChannel2|22.1.1.1/24": {}}, "INTERFACE": {"Ethernet12": {}, "Ethernet8": {}, "Ethernet12|2000:4::1/124": {}, "Ethernet12|24.1.1.1/24": {}, "Ethernet8|2000:3::1/124": {}, "Ethernet8|23.1.1.1/24": {}}, "PORTCHANNEL_MEMBER": {"PortChannel1|Ethernet0": {}, "PortChannel2|Ethernet4": {}}, "PORTCHANNEL": {"PortChannel1": {"admin_status": "up", "lacp_key": "auto", "min_links": "1", "mtu": "9100"}, "PortChannel2": {"admin_status": "up", "lacp_key": "auto", "min_links": "1", "mtu": "9100"}}}}}, "inconsistent_config": {"sonic-s6100-dut": {"null": {}}}}}
19:21:06 conftest.__dut_reload                    L1984 INFO   | dut reload called on sonic-s6100-dut
19:21:08 config_reload.config_reload              L0093 INFO   | reloading config_db
19:23:43 parallel.on_terminate                    L0085 INFO   | process __dut_reload--<MultiAsicSonicHost sonic-s6100-dut> terminated with exit code None
19:23:43 parallel.parallel_run                    L0221 INFO   | Completed running processes for target "__dut_reload" in 0:02:36.468107 seconds


========================================================================================================================== warnings summary ===========================================================================================================================
../../../../usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755
  /usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755: PytestAssertRewriteWarning: Module already imported so cannot be rewritten: tests.common.plugins.loganalyzer
    self.import_plugin(import_spec)

../../../../usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755
  /usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755: PytestAssertRewriteWarning: Module already imported so cannot be rewritten: tests.common.plugins.sanity_check
    self.import_plugin(import_spec)

../../../../usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755
  /usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755: PytestAssertRewriteWarning: Module already imported so cannot be rewritten: tests.common.plugins.test_completeness
    self.import_plugin(import_spec)

../../../../usr/local/lib/python3.8/dist-packages/_pytest/config/__init__.py:755
…